4.20-4.22

+++ part4/controllers/blogs.js
@@ -4,14 +4,6 @@ const Blog = require('../models/blog');
 const User = require('../models/user');
 const { SECRET } = require('../utils/config');
 
-const getToken = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   res.status(200).send(blogs);
@@ -29,12 +21,9 @@ blogRouter.get('/:id', async (req, res) => {
 
 blogRouter.post('/', async (req, res) => {
   const { body } = req;
-  const token = getToken(req);
-  const decodedToken = jwt.verify(token, SECRET);
+  const { token } = req;
 
-  if (!token || !decodedToken) {
-    return res.status(401).json({ error: 'token missing or invalid' });
-  }
+  const decodedToken = jwt.verify(token, SECRET);
 
   const user = await User.findById(decodedToken.id);
 
@@ -55,8 +44,15 @@ blogRouter.put('/:id', async (req, res) => {
 
 blogRouter.delete('/:id', async (req, res) => {
   const { id } = req.params;
-  await Blog.findByIdAndDelete(id);
-  res.status(204).end();
+  const { token } = req;
+  const decodedToken = jwt.verify(token, SECRET);
+  const blog = await Blog.findById(id);
+  if (blog && blog.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndDelete(id);
+    res.status(204).end();
+  } else {
+    res.status(401).send({ error: 'Not authorized' });
+  }
 });
 
 module.exports = blogRouter;
